Finalize build logic cleanup
