+++ build-logic/convention/src/main/kotlin/com/buzbuz/gradle/convention/libs/VersionCatalogWrapper.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 Kevin Buzeau
+ * Copyright (C) 2025 Kevin Buzeau
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,20 +14,18 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+package com.buzbuz.gradle.convention.libs
 
-dependencyResolutionManagement {
-    repositories {
-        google()
-        mavenCentral()
-    }
-    versionCatalogs {
-        create("libs") {
-            from(files("../gradle/libs.versions.toml"))
-        }
-    }
+import org.gradle.api.artifacts.MinimalExternalModuleDependency
+import org.gradle.api.artifacts.VersionCatalog
+import org.gradle.api.provider.Provider
+
+
+internal class VersionCatalogWrapper internal constructor(private val libs: VersionCatalog) {
+
+    val plugins: Plugins by lazy { Plugins(libs) }
+    val versions: Versions by lazy { Versions(libs) }
+
+    fun getLibrary(alias: String): Provider<MinimalExternalModuleDependency> =
+        libs.findLibrary(alias).get()
 }
-
-rootProject.name = "build-logic"
-include(":obfuscation")
-include(":convention")
-include(":source-download")
