+++ build-logic/convention/src/main/kotlin/com/buzbuz/gradle/convention/plugins/AndroidLocalTestConventionPlugin.kt
@@ -1,0 +1,50 @@
+/*
+ * Copyright (C) 2024 Kevin Buzeau
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.buzbuz.gradle.convention.plugins
+
+import com.buzbuz.gradle.convention.extensions.getLibs
+import com.buzbuz.gradle.convention.extensions.android
+import com.buzbuz.gradle.convention.extensions.androidTestImplementation
+import com.buzbuz.gradle.convention.extensions.plugins
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.dependencies
+
+class AndroidLocalTestConventionPlugin : Plugin<Project> {
+
+    override fun apply(target: Project): Unit = with(target) {
+        val libs = getLibs()
+
+        plugins {
+            apply(libs.plugins.jetBrains.kotlin.android)
+        }
+
+        android {
+            defaultConfig.testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+            testOptions.unitTests.isIncludeAndroidResources = true
+        }
+
+        dependencies {
+            androidTestImplementation(libs.getLibrary("junit"))
+            androidTestImplementation(libs.getLibrary("androidx.test.core"))
+            androidTestImplementation(libs.getLibrary("androidx.test.ext.junit"))
+            androidTestImplementation(libs.getLibrary("androidx.test.runner"))
+            androidTestImplementation(libs.getLibrary("androidx.test.rules"))
+        }
+    }
+}
