--- build-logic/core/src/main/java/com/buzbuz/gradle/core/libs/VersionCatalogWrapper.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2025 Kevin Buzeau
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.buzbuz.gradle.core.libs
-
-import org.gradle.api.Project
-import org.gradle.api.artifacts.MinimalExternalModuleDependency
-import org.gradle.api.artifacts.VersionCatalog
-import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.api.provider.Provider
-import org.gradle.kotlin.dsl.getByType
-
-
-class VersionCatalogWrapper internal constructor(private val libs: VersionCatalog) {
-
-    val plugins: Plugins by lazy { Plugins(libs) }
-    val versions: Versions by lazy { Versions(libs) }
-
-    fun getLibrary(alias: String): Provider<MinimalExternalModuleDependency> =
-        libs.findLibrary(alias).get()
-}
-
-
-fun Project.getLibs(): VersionCatalogWrapper =
-    VersionCatalogWrapper(extensions.getByType<VersionCatalogsExtension>().named("libs"))
