+++ build-logic/obfuscation/build.gradle.kts
@@ -41,7 +41,6 @@ dependencies {
     compileOnly(libs.androidx.room.gradlePlugin)
     compileOnly(libs.google.firebase.crashlytics.gradlePlugin)
     compileOnly(libs.google.gms.gradlePlugin)
-    implementation(project(":core"))
 }
 
 tasks {
