--- build-logic/parameters/src/main/kotlin/com/buzbuz/gradle/parameters/BuildParametersPlugin.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2025 Kevin Buzeau
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.buzbuz.gradle.parameters
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.create
-
-
-class BuildParametersPlugin : Plugin<Project> {
-
-    override fun apply(target: Project): Unit = with(target) {
-        extensions.create<BuildParametersPluginExtension>(PLUGIN_EXTENSION_NAME).apply {
-            rootProject.set(target.rootProject)
-        }
-    }
-
-    private companion object {
-        const val PLUGIN_EXTENSION_NAME = "buildParameters"
-    }
-}
-
